+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## What's Changed
 
+- Add create_backup_and_download to BackupProvider/CentralUnit
 - Refactor Client
 
 # Version 2025.12.2 (2025-12-01)